DM-9944 Use Plotly to create scatter version of the charts - basic scatter plot functionality (zoom, highlight, select)
